+++ insteon_mqtt/device/Switch.py
@@ -5,6 +5,7 @@
 #===========================================================================
 import functools
 from .Base import Base
+from . import functions
 from ..CommandSeq import CommandSeq
 from .. import handler
 from .. import log
@@ -17,7 +18,7 @@ LOG = log.get_logger()
 
 
 #===========================================================================
-class Switch(Base):
+class Switch(functions.Scene, Base):
     """Insteon on/off switch device.
 
     This class can be used to model any device that acts like a on/off switch
@@ -62,14 +63,8 @@ class Switch(Base):
             'on' : self.on,
             'off' : self.off,
             'set' : self.set,
-            'scene' : self.scene,
             'set_flags' : self.set_flags,
             })
-
-        # Special callback to run when receiving a broadcast clean up.  See
-        # scene() for details.
-        self.broadcast_done = None
-        self.broadcast_reason = ""
 
         # Update the group map with the groups to be paired and the handler
         # for broadcast messages from this group
@@ -189,59 +184,6 @@ class Switch(Base):
             self.off(group, mode, reason, on_done)
 
     #-----------------------------------------------------------------------
-    def scene(self, is_on, group=0x01, reason="", on_done=None):
-        """Trigger a scene on the device.
-
-        Triggering a scene is the same as simulating a button press on the
-        device.  It will change the state of the device and notify responders
-        that are linked ot the device to be updated.
-
-        Args:
-          is_on (bool):  True for an on command, False for an off command.
-          group (int):  The group on the device to simulate.  For this device,
-                this must be 1.
-          reason (str):  This is optional and is used to identify why the
-                 command was sent. It is passed through to the output signal
-                 when the state changes - nothing else is done with it.
-          on_done: Finished callback.  This is called when the command has
-                   completed.  Signature is: on_done(success, msg, data)
-        """
-        LOG.info("Switch %s scene %s", self.addr, "on" if is_on else "off")
-        assert group == 0x01
-
-        # Send an 0x30 all link command to simulate the button being pressed
-        # on the switch.  See page 163 of insteon dev guide
-        cmd1 = 0x11 if is_on else 0x13
-        data = bytes([
-            group,  # D1 = group (button)
-            0x00,   # D2 = use level in scene db
-            0x00,   # D3 = on level if D2=0x01
-            cmd1,   # D4 = cmd1 to send
-            0x00,   # D5 = cmd2 to send
-            0x00,   # D6 = use ramp rate in scene db
-            ] + [0x00] * 8)
-        msg = Msg.OutExtended.direct(self.addr, 0x30, 0x00, data)
-
-        # Use the standard command handler which will notify us when the
-        # command is ACK'ed.
-        done_callback = on_done if is_on else None
-        callback = functools.partial(self.handle_scene, reason=reason)
-        msg_handler = handler.StandardCmd(msg, callback, done_callback)
-        self.send(msg, msg_handler)
-
-        # Scene triggering will not turn the device off (no idea why), so we
-        # have to send an explicit off command to do that.  If this is None,
-        # we're triggering a scene and so should bypass the normal
-        # handle_broadcast logic to take this case into account.  Note that
-        # if we sent and off command either before or after the ACK of the
-        # command above, it doesn't work - we have to wait until the
-        # broadcast msg is finished.
-        if not is_on:
-            self.broadcast_done = \
-                functools.partial(self.off, group=group, on_done=on_done,
-                                  reason=reason)
-
-    #-----------------------------------------------------------------------
     def set_backlight(self, level, on_done=None):
         """Set the device backlight level.
 
@@ -359,10 +301,6 @@ class Switch(Base):
         # weird special case - see scene() for details.
         if msg.cmd1 == Msg.CmdType.LINK_CLEANUP_REPORT:
             LOG.info("Switch %s broadcast ACK grp: %s", self.addr, msg.group)
-
-            if self.broadcast_done:
-                self.broadcast_done()
-            self.broadcast_done = None
             return
 
         # On/off command codes.
@@ -375,10 +313,7 @@ class Switch(Base):
             if is_on:
                 self._set_is_on(True, mode, reason)
 
-            # For an off command, we need to see if broadcast_done is active.
-            # This is a generated broadcast and we need to manually turn the
-            # device off so don't update it's state until that occurs.
-            elif not self.broadcast_done:
+            else:
                 self._set_is_on(False, mode, reason)
 
         # Starting or stopping manual mode.
@@ -446,35 +381,6 @@ class Switch(Base):
         self._set_is_on(is_on, mode, reason)
         on_done(True, "Switch state updated to on=%s" % self._is_on,
                 self._is_on)
-
-    #-----------------------------------------------------------------------
-    def handle_scene(self, msg, on_done, reason=""):
-        """Callback for scene simulation commanded messages.
-
-        This callback is run when we get a reply back from triggering a scene
-        on the device.  If the command was ACK'ed, we know it worked.  The
-        device will then send out standard broadcast messages which will
-        trigger other updates for the scene devices.
-
-        Args:
-          msg (message.InpStandard): The reply message from the device.
-          on_done: Finished callback.  This is called when the command has
-                   completed.  Signature is: on_done(success, msg, data)
-          reason (str):  This is optional and is used to identify why the
-                 command was sent. It is passed through to the output signal
-                 when the state changes - nothing else is done with it.
-        """
-        # Call the callback.  We don't change state here - the device will
-        # send a regular broadcast message which will run handle_broadcast
-        # which will then update the state.
-        LOG.debug("Switch %s ACK: %s", self.addr, msg)
-
-        # Reason is device because we're simulating a button press.  We
-        # can't really pass this around because we just get a broadcast
-        # message later from the device.  So we set a temporary variable
-        # here and use it in handle_broadcast() to output the reason.
-        self.broadcast_reason = reason if reason else on_off.REASON_DEVICE
-        on_done(True, "Scene triggered", None)
 
     #-----------------------------------------------------------------------
     def handle_group_cmd(self, addr, msg):