+++ insteon_mqtt/mqtt/EZIO4O.py
@@ -20,8 +20,7 @@ class EZIO4O:
     allow input MQTT messages to change the state of the device.
 
     EZIO4O will report their state (state/output) and can be commanded to turn
-    on and off (set/output topic) or trigger a device scene with which the
-    modem is a responder (scene/output topic).
+    on and off (set/output topic).
     """
 
     def __init__(self, mqtt, device):
@@ -46,12 +45,6 @@ class EZIO4O:
             payload='{ "cmd" : "{{value.lower()}}" }',
         )
 
-        # Input scene on/off command template.
-        self.msg_scene = MsgTemplate(
-            topic="insteon/{{address}}/scene/{{button}}",
-            payload='{ "cmd" : "{{value.lower()}}" }',
-        )
-
         device.signal_on_off.connect(self._insteon_on_off)
 
     #-----------------------------------------------------------------------
@@ -70,7 +63,6 @@ class EZIO4O:
         self.msg_state.load_config(data, "state_topic", "state_payload", qos)
         self.msg_on_off.load_config(data, "on_off_topic",
                                     "on_off_payload", qos)
-        self.msg_scene.load_config(data, "scene_topic", "scene_payload", qos)
 
     #-----------------------------------------------------------------------
     def subscribe(self, link, qos):
@@ -93,10 +85,6 @@ class EZIO4O:
             topic = self.msg_on_off.render_topic(data)
             link.subscribe(topic, qos, handler)
 
-            handler = functools.partial(self._input_scene, group=group)
-            topic = self.msg_scene.render_topic(data)
-            link.subscribe(topic, qos, handler)
-
     #-----------------------------------------------------------------------
     def unsubscribe(self, link):
         """Unsubscribe to any MQTT topics the object was subscribed to.
@@ -108,9 +96,6 @@ class EZIO4O:
             data = self.template_data(button=group)
 
             topic = self.msg_on_off.render_topic(data)
-            link.unsubscribe(topic)
-
-            topic = self.msg_scene.render_topic(data)
             link.unsubscribe(topic)
 
     #-----------------------------------------------------------------------
@@ -204,35 +189,5 @@ class EZIO4O:
             self.device.set(level=is_on, group=group, mode=mode, reason=reason)
         except:
             LOG.error("Invalid EZIO4O on/off command: %s", data)
-
-    #-----------------------------------------------------------------------
-    def _input_scene(self, client, data, message, group):
-        """Handle an input scene MQTT message.
-
-        This is called when we receive a message on the scene trigger MQTT
-        topic subscription.  Parse the message and pass the command to the
-        Insteon device.
-
-        Args:
-          client (paho.Client):  The paho mqtt client (self.link).
-          data:  Optional user data (unused).
-          message:  MQTT message - has attrs: topic, payload, qos, retain.
-        """
-        LOG.debug(
-            "EZIO4O output %s message %s %s", group, message.topic,
-            message.payload
-        )
-
-        # Parse the input MQTT message.
-        data = self.msg_scene.to_json(message.payload)
-        LOG.info("EZIO4O input command: %s", data)
-
-        try:
-            # Scenes don't support modes so don't parse that element.
-            is_on = util.parse_on_off(data, have_mode=False)
-            reason = data.get("reason", "")
-            self.device.scene(is_on, group, reason)
-        except:
-            LOG.error("Invalid EZIO4O scene command: %s", data)
 
     #-----------------------------------------------------------------------